% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setEbayToken}
\alias{setEbayToken}
\title{Search ebay}
\usage{
setEbayToken(token)
}
\arguments{
\item{token}{Valid ebay developers token}
}
\description{
Set your personal ebay developers token as an enviroment variable, so you don't
have to specify it in each \code{search_ebay} call
}
\examples{
setEbayToken("YOUR_TOKEN")
}
